Move pages from app to components
