+++ components/resources/ResourceDetailCard.tsx
